Work-around for performance problems on ROOT 5.26.00, patch level a and later. If you have performance issues, please run with --autoflush 0.  Another option --autosave was introduced.  Both options take default values from the ROOT TTree implementation.
In ROOT 5.26.00-a new functionality was introduced in TTree to enforce basket
size optimization (TTree::OptimizeBaskets).  This seems to clash with our code,
or with how we are filling the branches.  The basket size optimization is done
when autoflushing of baskets to disk, and the first optimization happens when
the file size given by fAutoFlush is reached.  By default this is 30Mb (which
is different from the file size on disk becaise of the histograms and the other
tree).

By setting fAutoFlush to zero, baskets are only flushed when fAutoSave is reached
and no optimization based on fAutoSave is attempted.

For more details, this is the ROOT::TTree diff that caused the problem:
  http://root.cern.ch/viewcvs/trunk/tree/tree/src/TTree.cxx?r1=31941&r2=31960


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1452 b9700a98-9937-0410-b777-f318d6eb5bff