There was another mix-up of U and V in the tracking.  This time it was propagated by the uv2xy transformation helper.  Also fixed the angles to use the Qweak units.
The geometry file was updated with these consistent angle assignments.  Region 2 is
x location is not clear.  Region 3 has a V plane first, so the order is V U V' U'.
Doesn't matter for the tracking code.  What is more complicated is that the left
package in region 3 is not a rotation of the right package... but that can also
be dealt with.

Region 2 did not have a necessary local coordinate transformation in the transition
from Geant4 MC file to Qweak local octant coordinates.  And the wire offset is now
properly taken into account, as opposed to assuming that the central wires just
cross in one point.

Compiler warning when calculating cerenkov light yield response fixed.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1470 b9700a98-9937-0410-b777-f318d6eb5bff