Add some extra scaler channel decoding to the main detector map.
It looks like the main detector scaler actually WAS running during
last weekend's data, but the cable was swapped around and possibly
upside down.  This commit just decodes the remaining scaler channels
with offensive names and we can sort out what was actually the case 
later.

The naming scheme corresponds to a possible way to refer to mapfiles 
from old runs, blame Paul.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1477 b9700a98-9937-0410-b777-f318d6eb5bff