This is what I've been using to analyze the scaler readouts for tracking mode runs.  I run qwparity on a tracking run and ignore everything except the Mps_Tree leaves I'm interested in.  s0ch29 should be 31 MHz clock input, s0ch30 is a 4 MHz clock, and s0ch31 should be 31 MHz return.  For some runs s0ch29 and s0ch31 are 14 MHz...?
When a tracking trigger and a parity trigger happen at the "same" time the parity scaler fills the buffer normally but it is not readout and cleared.  So the next parity trigger tries to read the buffer normally but now there are two events in the buffer.  I made a check to see if this is the case.  If so, it just drops the events, sets them to zero and moves on.

This is not the best way to do this but I have many other things to worry about right now.

 





git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1494 b9700a98-9937-0410-b777-f318d6eb5bff