Changes to the blinder and the helicity pattern class to pass the EPICS and parity event.  This is then used to determine whether/how to blind the data. The default (before there is any EPICS or parity DAQ data) is to not blind, but this is reevaluated for every EPICS event and for every helicity pattern.
Currently this just includes the calls, and it just checks whether the BCMs
are seeing something.  If they do, then attempt to read seed_id 1 from the
database.  I would like to write something like a channel map to be able to
change the conditions easily, but somehow that is maybe not such a good idea...

Possible issues with checking so often is that we could be overloading the
database with requests (this runs offline from the farm, possibly in parallel).

Expect to see changes to this before this goes 'live'.  Until then, blinding
remains disabled by default.

Also, changes to the EPICS event to require const-correctness.  QwEPICSEvent
should really use a (hash)map to do name-to-index lookup instead of looping
over all entries and doing string comparisons...


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1504 b9700a98-9937-0410-b777-f318d6eb5bff