Added new member function QwBlinder::ReadSeed(QwDatabase*) that fetches the correct seed (and seed_id) for a given run based on its run number.  In version 01.00.0004 of the database schema the seeds table was upgraded to store the run range a particular seed is valid for.  Setting the seed_id manually, hence use of QwBlinder::ReadSeed(QwDatabase*, UInt_t), should be deprecated.
New database installations come with a default seed installed in the seeds table that covers the default run range 0 to 1000000.  New seeds can be inserted (with corresponding updates to valid run range) using the insert_seed.pl script in the QwDB_Admin module.

I simply added this function to the class; I did not update the calling functions to use this version rather than the original version with manually specified seed_id.  Not sure how this new approach fits into Wouter's scheme.

With this commit, QwDatabase will only connect to databases implementing schema 01.00.0004 and greater.  There is a test database on cdaql6 that implements this version.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1512 b9700a98-9937-0410-b777-f318d6eb5bff