Tree trimming is now available for the offline parity analyzer. There is a default tree trimming file called "Qweak_Tree_Trim_List.in" that is loaded by the QwParity executable. When ever a branch is created at the module level, the branch base name is checked against the tree trim list, and if the name exist in the list, then the branch is created.
The tree trimming can be enabled if --enable-tree-trim flag is applied at the command line.

The current version of the Qweak_Tree_Trim_List.in contains all the branch names currently available in the analyzer, So even with the trim tree enabled the entire tree will be created. One can comment out (use #) unwanted devices, improve the analyzer performance. 

The Qweak_Tree_Trim_List.in list also supports the similar regular expressions used in the histogram definition file.
This is a step towards a better tree trimming capability in the offline analyzer. Comments and improvements to the scheme are always important.



If you want to generate a list of branch names in your own qweak trees, there is a app for that!!!! actually a script in the ~/Extensions/Macro/Parity/ListBranches.C

that has two routines
	void ListBranches(int run_number)
	void ListBranches(TString Rfilename)
And will generate a text file containing list of branch names. See the script for more information.

More information about the trimmed tree implementation can be found at

https://qweak.jlab.org/wiki/index.php/DAQ_Analyzer#Trimmed_Tree_Implementation

Rakitha



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1515 b9700a98-9937-0410-b777-f318d6eb5bff