Update the trig scint channel map: -Definitions of 1 and 2 have been switched to be consistent with package numbering of chambers (package 1 is beam rest when in the default/rest position) -After changing the cabling at the splitter panel the ADC channels were wrong -Change naming format to be more consistent with main detectors -Added the scaler channels
However, while the scalers are getting data (seen in xcefdmp), they are not being read out yet.  I made the appropriate modifications to the trigger scintillator class to handle the scaler data.  It seems to work because I can fill scaler histograms with main detector values.  However, I cannot fill the tree with trigger scint scalers yet.  The only difference between these is the module number, which is 0 for the TS and 1 for the MD.  I haven't looked at this in detail yet, but I will do some debugging to see what's going on.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1545 b9700a98-9937-0410-b777-f318d6eb5bff