Fixed the problem that missing pattern in the helicity tree
The Fix and the Cause of the Error

Added new boolean parameter to store the Ring status in the QwParity main routine. Calling the QwEventRing::IsReady() multiple times within the same event sometimes returns indifferent state (i.e returning true and false in the within an event). This was the reason behind some events are not going into the helicity pattern tree. The only way to prevent this is to call QwEventRing::IsReady() once  just after QwEventRing::push(detectors) and save the sate in a boolean parameter. Then use this value in the rest of the code. This has been the case in QwAnalysis_BeamLine executable but in the QwParity executable, QwEventRing::IsReady() is called in multiple instances. I have fixed this now.

Rakitha


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1546 b9700a98-9937-0410-b777-f318d6eb5bff