Add a histogram that plots the package values (either 1 or 2)  of Region 2 and Region 3 when an event has both Region 2 and Region 3 data.
Currently the histogram plots the four possible combinations of 
package values (11,12,21,22)  with the number in the tens place value 
indicating the package value for the region 2 data  and the ones places 
value indicating the region 3 package value. (e.g. an entry in the 
histogram of '21' indicates that there was an event where the Region 2 
package value was '2' and the Region 3 package value was '1')

 


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1655 b9700a98-9937-0410-b777-f318d6eb5bff