EventDisplay: The rootfile structure should now be consistent with udpated root trees. (Notice: This breaks compatibility with older rootfiles.)
Added the ability to roughly draw partial tracks. Though, this section still needs a lot of work depending on the progress of the tracking. At the very least, this is a foundation upon which to build the complete event display tracking.

GNUmakefile: Added the missing debug option into the compiler.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1703 b9700a98-9937-0410-b777-f318d6eb5bff