Note:  This is on the Moller polarimeter development branch.
Cleanup of the Moller branch in preparation for the merge into the
trunk; removed a few extra output comments, and removed some 
meaningless differences w.r.t. the trunk.
Modified the Helicity class to allow the H+ and H- event types to be
specified from the map file for the Moller helicity mode.
Shifted some code from "QwHelicity::ProcessEvent" into the Moller-specific
helicity processing method to avoid conflicts with the main analyzer.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/moller@1706 b9700a98-9937-0410-b777-f318d6eb5bff