Merged the Moller polarimeter development branch into the trunk using:        svn merge https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1702        	   https://qweaksvn.jlab.org/repos/QwAnalysis/branches/moller@1706 .
  Removed the unneeded executable:
  	Parity/main/QwAnalysis.cc

  Reverted unneeded differences in:
	Analysis/include/QwEventBuffer.h
	Analysis/src/QwHistogramHelper.cc
	Analysis/src/QwVQWK_Channel.cc
	Parity/src/QwSubsystemArrayParity.cc
	Parity/src/QwHelicityPattern.cc
	Parity/prminput/qweak_helicity.map

  The Moller analyzer results have been compared against the previous
versions which were tested against the FORTRAN analyzer.

  The Moller analyzer is hardcoded to process data file 15; the analyzer
will require substantial modifications to use the most up-to-date versions
of the Parity analysis support classes (it was based on an old version of
QwAnalysis_Beamline.cc instead of QwParity.cc).

NOTE:  during all of this, the local svn record for the GNUmakefile
got corrupted compared to the repository.  I reverted it to the
trunk version (without the Moller).
This version WILL NOT include the QwMoller in the executable targets
by default.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1707 b9700a98-9937-0410-b777-f318d6eb5bff