Update to the units and geometry used in the tracking. This improves the number of good region 3 tracks found.
A while ago we agreed on mm as the basic unit to be used
by anyone.  The result is that we now have a couple of
places in the code where we dutifully transform cm in mm,
pass it to a function, and in the function tranform mm to
cm.  Since everyone is using cm it is just easier to change
the convention:

  Please use cm as the unit of length in QwTracking

The geometry file (which was always in cm) was also modified
a bit to improve reconstruction of tracks in the second
package kPackageDown (i.e. from 0 tracks to a reasonable
number of partial tracks found).


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1716 b9700a98-9937-0410-b777-f318d6eb5bff