Below follows a first attempt at a commit log message that I typed last week.  It didn't go through and a bunch of it has changed but I don't feel like rewriting all of it, so here you go.  Note: prescaling and tree disable flags work now. --
Changes to the way ROOT files are written.  No changes in ROOT files.

Until now ROOT files were written using QwRootFile only in parity mode.
In tracking mode we used hardcoded calls to TFile and TTree.  I rewrote
some of the QwRootFile class so it does not depend on parity-specific
code anymore (and can now rightfully continue to live in the Analysis
directory).  The options are now parsed in QwOptions.cc instead of the
include QwOptionsParity.h.

QwRootFile now also defines a QwRootTree.  This is a helper class that
includes a set branches and a value vector.  It includes a reference
to a TTree, so multiple QwRootTree objects can be in one TTree.  The
calling syntax for QwRootFile output is slightly different, and you
have to specify by string which tree you want to write to:

    rootfile = new QwRootFile(eventbuffer.GetRunLabel());
    rootfile->ConstructHistograms("mps_histo", detectors);
    rootfile->ConstructHistograms("hel_histo", helicitypattern);
    rootfile->ConstructTreeBranches("Mps_Tree", "MPS event data tree", detectors);
    rootfile->ConstructTreeBranches("Hel_Tree", "Helicity event data tree", helicitypattern);
    ...
        rootfile->FillHistograms(detectors);
        rootfile->FillTreeBranches("Mps_Tree", detectors);
        ...
              rootfile->FillHistograms(helicitypattern);
              rootfile->FillTreeBranches("Hel_Tree", helicitypattern);

This is all using template functions where class T has to provide
  T::ConstructBranchAndVector
  T::ConstructHistograms
  T::FillTreeBranches
This is the case for QwSubsystemArray and QwHelicityPattern.

The reason for this is that there might be a situation where we have

        rootfile->FillTreeBranches("Mps_Tree", detectors_before_regression);
        rootfile->FillTreeBranches("Mps_Tree2", detectors_after_regression);

Both detectors_before_regression and detectors_after_regression have
exactly the same type, so you have to give some more information to
make sure that the data is written to a different tree (or different
branches).  But this is still somewhat in flux, so that could change
slightly.  The type info is used to prevent

        rootfile->FillTreeBranches("Mps_Tree", helicitypattern);

but that is not a complete solution.  We could store the pointer address
of the object used to construct the branches and compare that when we
fill the tree, but that seems wrong somehow.  Constructing with one
object and filling with a copy should be possible.

In the process I had to disable tree prescaling (until that can be
generalized to the new lists of trees), and the flags --disable-mps
and --disable-hel.  I don't think anyone is really using that now,
but if you are you shouldn't update yet.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1743 b9700a98-9937-0410-b777-f318d6eb5bff