The blinder will now use EPICS records to determine if a blindable target is in position (a later update will allow a word in the event data to indicate this). If a blindable target is in position, and the beam current is above a specified threshold, then the blinding term will be applied to the main detector channels; if the beam current is not above the threshold, the main detector asymmetries and differences will be marked as in error, and the main detector asymmetries and differences will be set to zero for that helicity pattern. If a not-blindable target is in position, then no asymmetries will be cut for any beam current. If the target position cannot be determined, then the blinder will fail the blindable main detector asymmetries.
Individual PMT channels and combined PMT channels in the main detector
can be marked as not-blindable, and these channels will not have the
asymmetries cut.

The "alternate asymmetries" are never blinded.

There are command line flags to force the target position to be interpreted
as either blindable or not-blindable, as well as to alter the beam current
threshold.

If the target position changes during the run, the "blindablility" of
the target which was set first will be used throughout, and the data
in the second "blindablility" will have the blindable main detector
asymmetries be failed.

All of the blinding is done on the differences before the asymmetry
is calculated.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1797 b9700a98-9937-0410-b777-f318d6eb5bff