Added sensitivity scaling factors and relative gains for the bpms in to the qweak_beamline_geometry.map file. h. The sensitivity scaling factor is the ratio between actual BPM sensitivity (previously defined in the QwBPMStripline as kQwStriplineCalibration) and the default BPM sensitivity. The  default BPM sensitivity (BSEN) for many of the BPMs in the hall C and injector beamline is 18.81 adc counts/mm. But there are few BPMs which have different sensitivities (e.g. 3H08 BSEN = 25.8) . To enable use of different sensitivity factors, the above mentioned ratio was added in to the geometry file. The relative gains (or the XALP/ YALP in epics) for the X and Y wires of a BPM take into account the gain difference between the + wire and - wire.  Using these, modifies the RelX and RelY position determination in the BPM frame to
RelPosX = (XP - AlphaX*XM)/(XP + AlphaX * XM) 

and similarly for RelPosY. 

kQwStriplineCalibration = 18.81*sensitivity scaling factor.

All of these values obtained from EPICs are now listed in the geometry file. We need to use all of this factors to make sure our readings from the BPMs are more accurate. Changes were tested with run 5083

The change in the VqwSubsystem files are related to the name change I did for the LoadGeomentryParameters()  function in the QwBPMStripline class. Earlier it was called LoadGeomentryDefinitions().

Added a flag for a debugging line in the EPICSEvent class which was printing all the epics variables. I was not sure if this was really needed so I put a flag and made it FALSE for now.

All the other changes are related to the implementation of the scaling factors and relative gains for the QwBPMStripline class.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1798 b9700a98-9937-0410-b777-f318d6eb5bff