This commit does two things, both related to time-normalization.
The first is that the VQWK channel ProcessEvent will divide the
hardware sum and sub-block values by the number of samples in
each.  The results for the hardware sum and for the sub-blocks
are now the average single-sample values.
An option to convert the reported values to voltage will be
provided soon.

The second is that all helicity pattern yields and differences
have been divided by the number of helicity windows in the pattern.
Yields are now the average yield for a single helicity window,
and the differences are the average deviation from the mean,
such that:
     Mean_H+_state_value = Mean_yield + Mean_difference
     Mean_H-_state_value = Mean_yield - Mean_difference
In this scheme, the pair-differences are twice the reported
mean difference.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1870 b9700a98-9937-0410-b777-f318d6eb5bff