Important notice: this commit could break macros!
Removed the hit_tree from the tracking ROOT file.  The same information is
very easily accessible from the event_tree.  This will avoid writing the hits
three times to the tracking ROOT file.  If you were using the syntax
  hit_tree->Draw("hits.fQwHits.fTime")
then replace that with
  event_tree->Draw("events.fQwHits.fTime")

Also the branch hits in the event_tree was removed.  If you were using
  event_tree->Draw("hits.fQwHits.fTime")
then replace that with
  event_tree->Draw("events.fQwHits.fTime")

If you don't feel like doing that right now, don't upgrade.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1875 b9700a98-9937-0410-b777-f318d6eb5bff