New features added to the Real Time GUI. 1. The real time GUI now reads the detector map file to access channel maps for individual subsystems to load the device lists. The flag '-d' can be passed to load a custom detector map at the time of staring the GUI by,
./QwRealTimeGUI -d detector_240Hz.map

The default ./QwRealTimeGUI will load the standard detector map file 'detectors.map'

GUI will add subsystem tabs only if the subsystem is included in the detector map file. So Real time GUI will show only tabs corresponding to subsystems that are loaded in the qwparity analyzer. 

2. The histogram definitions in the file 'Qweak_RT_Hist_Trim_List.in', upper and lower limits of all the histograms are set to the same value (=0). This will enable root to setup lower and upper limits based on the data filled into the histograms. root will scan data till maximum buffer size of entries and decides the limits. Currently the buffer size is set to 1000 entries. This can be adjusted in the 'Qweak_RT_Hist_Trim_List.in'
	DefaultBufferSize=1000

In the near future when we have good parity quality data we may set actual lower and upper limits to the histograms properly. But until then we will stick to the auto limits. 

3. For summary plots in all subsystems, mean value and the standard deviation is plotted for all the devices. So the error bars corresponding to the standard deviation of the mean.

Rakitha


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1883 b9700a98-9937-0410-b777-f318d6eb5bff