Updates to QwFeedback
Added upper and lower limits to the IA set-points into the feedback parameter file. These limits are used to verify the calculated IA set-points are within the limits. This will prevent setting IA DACs with unreasonable values calculated by the feedback algorithm when the charge asymmetries are too large (+/-).

Changes to the helicity mode determination, so that same convention is followed between the IA DAC and QwFeedback analyzer.

Rakitha



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1900 b9700a98-9937-0410-b777-f318d6eb5bff