!!! Merge of branch wdconinc.param_file_with_run_range into trunk. !!!
After consultation with Paul, the branch with run-specific parameter files
is now merged into trunk.  This means the following:

- The detectors.map should not contain run number-specific parameter files.
- Parameter files can have a run label of the format name.run1-run2.ext,
  or name.run1-.ext, or just name.run1.ext.  name.ext is also still valid,
  and is used in the absence of more specific parameter files.
- EPICS events are processed using the QwParameterFile class instead of the
  legacy StringManip functions.

For more information on what went into the branch, please take a look at the
branch changelog:
  svn log --stop-on-copy https://qweaksvn.jlab.org/repos/QwAnalysis/branches/wdconinc.param_file_with_run_range

If you notice unexpected behavior, check which file is being used by scrolling
through the screen output.  The used parameter file is always printed to the
screen.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1910 b9700a98-9937-0410-b777-f318d6eb5bff