Added channels to the analyzer for the new background detectors.
I added the new channels in the scanner system, so they will be
accessible from root files created with both the parity and tracking
analyzers.  The names are ugly.

The discriminator signals for the new detectors, like the three
scalers for the scanner, go both into the trigger supervisor and into
ROC4.  The ribbon cable into the trigger supervisor is flipped, so
that the ECL signals in those scalers will have the wrong polarity.
However this flip seems to be codified in the scanner map, where the
both sets of detectors are present.  I think this is an artifact from
the time before ROC4 was in the data stream, and so I've switched both
the scanner scaler signals and the new signals to look for the data
from ROC4.

git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1921 b9700a98-9937-0410-b777-f318d6eb5bff