There were complaints that it was annoying to have to wait for the main detector data to load into the GUI, when all one wanted was (say) the beam information.
So I modified the GUI, such that now, if you deselect
a tab (i.e. remove a tab) for a particlar subsystem,
from the display, the data for that subsystem will not
be read from the file.

So, for example, if you want to look at the hallc 
beamline tab, but NOT the main detector tab, you can
remove the main detector tab (using the "Tabs" menu)
and then open the file. This will then prevent the
main detector data from being loaded. 

If a file is already open and a tab is added to the
list of active tabs, then the data for that tab will
be loaded at that point.

By default all of the subsystem tabs are shown, so 
you have to remember to remove the subsystem tabs you
don't want before opening a file.

This currently applies to ROOT files only, not the
database files.

Hope this helps.

Michael





git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1938 b9700a98-9937-0410-b777-f318d6eb5bff