Added a few more little scripts for MD plots. All of these are readily available in a more robust manner in the GUI's, so they're not immensely useful.
~------------------------~

On Tuesday, Oct. 12th, 2010, we took some data at 120 Hz. None of our mapfiles were set to agree with this, so I've devised a set of mapfiles for the Main Detectors, Lumis, and Beamline that has the correct number of samples (=4000)

I've also added a new detector map called detectors_120.map that points to these files. So one simply needs to include the "--detectors detectors_120hz.map" flag when running qwparity.

git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1961 b9700a98-9937-0410-b777-f318d6eb5bff