Changed main detector pedestal file (and the pedestal script) to automatically scale from "counts" to Volts, for the main detector channels.
Updated the MD GUI to NOT do this scaling again. (Has anyone else had compilation issues with the Offline GUI lately? Mine won't compile...)

+----------------+

IMPORTANT UPDATE FOR TONIGHT!!!

I have, with Paul's blessing, enabled the event cuts in the qweak_beamline_eventcuts.in file and am cutting on bcm1. Any events that have bcm1 falling below 2 uA will be discarded from the ROOTFile. This was done so that the ROOTFile will play nicely with the linear regression code. On the flip side, this means that we will not have any 'pedestal data' during beam trips in these ROOTFiles. If you want to look at signals during a beam trip in the data you will have to reanalyze the run.

git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1963 b9700a98-9937-0410-b777-f318d6eb5bff