Removed a function related error in the QwGUI hallC tab.
Error in <TQObject::CheckConnectArgs>: slot DoHPosition(int) does not exist

This only appears when the GUI is started. Its not a compiler error. I was trying to use radio button groups to plot correlations and was trying to use signal slot communication mechanism in root. The error was related to that. But it was harmless as the function mentioned in that error, DoHPsition() was not being used by anything in the GUI. It was still under construction and was complaining. I should have removed it but I was going to fix it in my next commit which never happened. I did mention this in my commit though. Still people get comfused so I am removing that. Sorry for the comfusion.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1971 b9700a98-9937-0410-b777-f318d6eb5bff