This commit introduces support for run/event lists.  Not particularly useful for parity (someone will come up with a use), but certainly a useful debugging tool for tracking mode.  I guess it can also be used for batch processing if you don't want to use batchman on the farm.
The following file (specified with --runlist <filename>):

  [5253]
  234
  246
  256
  345:456
  567:789
  [5259]
  [5260]
  0:10000
  [5261:5270]
  9000:10000

will do what you expect it to do:
- for run 5253 it will analyze three individual events, and two event ranges
- for run 5259 it will analyze the entire run (all segments)
- for run 5260 it will analyze the first 10000 events
- for runs 5261 through 5270 it will analyze the events 9000 through 10000

The runlist is searched for in the standard search paths, which do not include the
pwd.  That might seem confusing at first (and also later).  Keep event ranges sorted
inside the run blocks.  Runs do not need to be ordered.

Also, VQWK channels will report division by zero when QwLog level is verbose.  It
happens fairly often naturally, so it is off by default so as not to confuse poor
shift workers.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2005 b9700a98-9937-0410-b777-f318d6eb5bff