Print something useful when the flag --QwTracking.showmatchingpattern is given.
For a region 3 event it will now print (with appropriate flags):

131:||....|||..||||
132:.||..||.|||||||
133:...||....||.|||
134:..|..|...||.|||
135:||....|||..||||
0x8f9518c: (2,*) 0,1,2,4,5,7,0,0, width = 8
|.......
.|......
..|.....
....|...
.....|..
.......|
|.......
|.......

The first 'graphic' is the hit pattern, then info about the tree line that was
found in this hit pattern.  I added the second 'graphic', which basically shows
what was in the pattern that matched.  Here, the pattern was at level 2, which
is the 3rd level, so with 8 bins.  The pattern had a hit in the indicated bins.
Each row is one wire, each column is one drift distance bin.

Main thing to keep in mind: top is hit pattern, with all ambiguities.  Bottom is
the idealized track that is consistent with that pattern (and there could be more
than one).  The bottom should look very clean.  Noisy hit patterns will just lead
to more treelines found, and thus many graphics of the second type.

Other tree objects were made more generic by not relying on #defined sizes for the
tree, but getting them from the geom file (via via).  This reduces the memory
occupied by the tree by a factor 2 (we weren't really worried about that), and
removes redundant info in debug outputs.

Printing the tree search db will not be recursive by default.

QwScaler had a ==/= typo.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2023 b9700a98-9937-0410-b777-f318d6eb5bff