A one step towards providing critical information using the RTGUI. Basic calculation of common mode noise is implemented in the Real Time (RT)GUI.
1. RTGUI Common Mode Noise Calculation
The common mode noise and fractional broadening are calculated in the RTGUI. The calculation is dependent on the RT flat tree. The main detector bar asymmetries must be in the tree trim file to calculate these properties.

I have used the Mark Pitt's suggestions to calculate common mode noise.
"Suggestion for quick estimate of the common mode noise" hclog entry 201563. 

The common mode noise is calculated in the Main detector tab. I have added a button "Common Mode Noise"

2. Limitations
Since the RT flat tree is by default circular buffered the statistics are limited and depend on the circular buffer. If we generate all necessary histograms in the RT engine, we could use histograms to compute these quantities. 
Currently NO hardware, beam or main detector related cuts are applied.

3. Trimmed the RT flat tree. 
So we could run the RT engine with the tree without compromising much of resources.
Tree Hel_Tree No.of Leafs 56 
Tree Mps_Tree No.of Leafs 26 

4. Increased the circular buffer size to 25000 (Parity/prminput/parity_realtime.conf)

Rakitha


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2054 b9700a98-9937-0410-b777-f318d6eb5bff