Get ready to change all of your macros and scripts, folks!
In an effort to standardize the naming in the qwparity analyzer, I have out the qwk_ prefix on all main detector channels, including the background detectors.

I have done this for the 3 mapfiles (.map, 120.map, 240.map) and retroactively for all of the pedestal files.

I have also changed the scaling factor in old pedestal values to reflect the conversion from ADC counts to voltages. This way if you're replaying an older run you can more easily compare it to a new one.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2083 b9700a98-9937-0410-b777-f318d6eb5bff