A script for identifying and killing open ssh tunnels.
One failure mode for the ssh_config commited to this directory
recently is that the ssh connection to the login server may stop
communicating with the server but not actually exit.  This makes new
connections fail mysteriously.

You can use

	$ tunnels

to list all the open ssh tunnels; or

	$ tunnels jlab

to list all the ssh tunnels whose command string matches "jlab" (which
string can be a Perl regular expression, if protected from the shell
by quoting); or

	$ tunnels jlab --kill

to sent a KILL signal to all the matching ssh processes.

Generally if my laptop is closed for a while, or I switch from one
network to another, a clean ssh slate with "tunnels --kill" makes
things go a little smoother.

git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2087 b9700a98-9937-0410-b777-f318d6eb5bff