Added the device weights for the individual bpms in the combination qwk_target. These weights were obtained using run 5832.I did a fit to each of the bpms using two of its negihbouring bpms and then used the rms of fit residual distribution to get the weights.
A big change (sort of) to the bpm list. Changed 3P01 to 3C07A.Why?
While looking at the 3C17 missing signals I found out that what we are reading as 3P01 in the DAQ is actually 3C07A. When we were pulling wires to these BPMs it was not clear at the running time if we'd have wither 3P01 or 3C07A. We ended up having 3C07A and I ended up doing all the labels properly in the ADCs and cables but I failed to change the name in the qweak_hallc_beamline.map file. I even had the right bpm (3C17) in the wiki and the pedestal files.
Long story short, all this time we were decoding 3C07A as 3P01 in to the rootfiles. Its not a hardware mix up it was a software mix up done by me.
Sorry..




git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2131 b9700a98-9937-0410-b777-f318d6eb5bff