Some updates and modifications to the beam on target projection script.
The values projected by this script using the same 7 bpms that the combined bpm qwk_target uses does not match with each other for obvious reasons. qwk_target uses least squares fit on event by event basis to get X and Y on target and hallc_beam_motion.cc uses the mean values of each bpm to project the beam on target using least squares fit.
But their values are within errors and to show this clearly I added an extra point with error bars to the graphs that contain the projected targetX and Y. 

The errors were calculated using error propergation method and I assumed that the error on the Z location was negligible. As you will see, the value from qwk_target bpm (green) falls well within the errors on value of target X from the fitted value obtained in this script(blue).





git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2143 b9700a98-9937-0410-b777-f318d6eb5bff