Small changes in the Compton photon detector class, mostly avoiding range checks when unnecessary.
Changes in the histogram helper class to speed up the start of the
analyzer.  Now, a regexp is created when the histogram file is read,
not when a histogram is created.  This search will always stay O(n^2)
which means that redundant lines in the histogram definition files
are really impacting (start-up) speed.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2149 b9700a98-9937-0410-b777-f318d6eb5bff