Created a script to access rootfiles within a user given run range to plot (for now) the mean and mean error of any signal that is in the histogram folders, as a function of the run number. Since its accessing the histograms and not the tree, I am not sure how to apply error checks like using DEvice_Error_Code == 0 in trees. Thats something that I have to think about later. I din't wanted to plot from the tree to get the mean and error because that takes a lot of time to do and it slows down the program.
The compiled script, read_run_range.cc creates the executable get_from_runs.
To use it, you have to use the command
 ./get_from_runs run1 run2 device property
where, run1 and run2 are the begining and end runs for the range,
       device is the device name as it appears in the device map files. (e.g. qwk_bpm3h09X)
       property can either be
          1. "a"  which is asymmetry in  hel_histo
          2. "d"  which is differences in hel_histo
          3. "y"  which is yield in hel_histo
          4. "s"  which is signal in mps_histo

I did my best to add proper units for currents and positions. But when you plot signals "s", don't be surprised if (e.g.) "qwk_bpm3h09XP" have mm displayed as units. Its not mm. Its adc counts and I have to fix that. I will take care of the cosmetics effects later.
 
Eventually, the work done by this macro will be something that will go in to the QwGUIDatabase tab in the Offline GUI. Till Damon fix the issues with beam summary tables filling, we can use this script to get an idea about the beam monitor readings during different running periods. Since MDs and Lumis can access there run based data from the QwGUIDatabase I didn't thought about plotting them specifically but in principle if one knows the correct name of the histogram in the histogram folders, one should be able to plot it using this script.

One small issue is that if this script finds a tracking rootfile while its scanning through the run range searching for the correct histogram, it will  ignore that root file (as it should do) saying it didnt find the histogram its looking for, but after like next 5+ runs the whole program will crash. I don't know whats causing this and it seems to happen only for tracking rootfiles. So have to be carefull of the run range you choose.
Other than that, I did check all possible crashing scenarios and put in proper error messages for that. If anyone finds any other sort of errors please let me know.

Next step will be to put this script in to work and to update it to use information like IHP state, event cuts, current cuts (?), etc.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2150 b9700a98-9937-0410-b777-f318d6eb5bff